+++ Playground/StackStatusTree/Input.h
@@ -5,5 +5,5 @@
 #include <utility>
 #include <vector>
 
-extern std::vector<std::pair<int64_t, int64_t>> gInput;
-extern std::unordered_map<int64_t, std::string> gDescriptions;
+extern std::vector<std::pair<uint32_t, uint32_t>> gInput;
+extern std::unordered_map<uint32_t, std::string> gDescriptions;
